Issue #SB-26318 feat: Moved the CloudStorageUtil class to jobs-core module and removed from asset-enrichment, auto-creator-v2 and publish-core module.
