+++ asset-enrichment/src/main/scala/org/sunbird/job/assetenricment/functions/VideoEnrichmentFunction.scala
@@ -7,9 +7,9 @@ import org.sunbird.job.assetenricment.domain.Event
 import org.sunbird.job.assetenricment.helpers.{OptimizerHelper, VideoEnrichmentHelper}
 import org.sunbird.job.assetenricment.models.Asset
 import org.sunbird.job.assetenricment.task.AssetEnrichmentConfig
-import org.sunbird.job.assetenricment.util.{CloudStorageUtil, YouTubeUtil}
+import org.sunbird.job.assetenricment.util.YouTubeUtil
 import org.sunbird.job.exception.InvalidEventException
-import org.sunbird.job.util.Neo4JUtil
+import org.sunbird.job.util.{CloudStorageUtil, Neo4JUtil}
 import org.sunbird.job.{BaseProcessFunction, Metrics}
 
 import scala.collection.JavaConverters._
@@ -31,6 +31,7 @@ class VideoEnrichmentFunction(config: AssetEnrichmentConfig,
   override def close(): Unit = {
     super.close()
   }
+
   @throws(classOf[InvalidEventException])
   override def processElement(event: Event, context: ProcessFunction[Event, String]#Context, metrics: Metrics): Unit = {
     logger.info(s"Received message for Video Enrichment for identifier : ${event.id}.")
