+++ asset-enrichment/src/main/scala/org/sunbird/job/assetenricment/helpers/ImageEnrichmentHelper.scala
@@ -4,13 +4,12 @@ import org.im4java.core.Info
 import org.slf4j.LoggerFactory
 import org.sunbird.job.assetenricment.models.Asset
 import org.sunbird.job.assetenricment.task.AssetEnrichmentConfig
-import org.sunbird.job.assetenricment.util.{AssetFileUtils, CloudStorageUtil, ImageResizerUtil}
+import org.sunbird.job.assetenricment.util.{AssetFileUtils, ImageResizerUtil}
 import org.sunbird.job.domain.`object`.DefinitionCache
-import org.sunbird.job.util.{Neo4JUtil, ScalaJsonUtil, Slug}
+import org.sunbird.job.util.{CloudStorageUtil, Neo4JUtil, ScalaJsonUtil, Slug}
 
 import java.io.File
 import scala.collection.mutable
-
 
 trait ImageEnrichmentHelper {
