+++ asset-enrichment/src/main/scala/org/sunbird/job/assetenricment/helpers/OptimizerHelper.scala
@@ -2,7 +2,7 @@ package org.sunbird.job.assetenricment.helpers
 
 import org.slf4j.LoggerFactory
 import org.sunbird.job.assetenricment.models.Asset
-import org.sunbird.job.assetenricment.util.CloudStorageUtil
+import org.sunbird.job.util.CloudStorageUtil
 
 import java.io.File
