+++ auto-creator-v2/src/main/scala/org/sunbird/job/autocreatorv2/util/FileUtils.scala
@@ -2,7 +2,7 @@ package org.sunbird.job.autocreatorv2.util
 
 import org.apache.commons.lang3.StringUtils
 import org.slf4j.LoggerFactory
-import org.sunbird.job.util.JSONUtil
+import org.sunbird.job.util.{CloudStorageUtil, JSONUtil}
 
 import java.io.{File, IOException}
 import java.net.URL
