+++ jobs-core/src/main/scala/org/sunbird/job/BaseJobConfig.scala
@@ -1,13 +1,13 @@
 package org.sunbird.job
 
-import java.io.Serializable
-import java.util.Properties
-
 import com.typesafe.config.Config
 import org.apache.flink.api.common.typeinfo.TypeInformation
 import org.apache.flink.api.java.typeutils.TypeExtractor
 import org.apache.kafka.clients.consumer.ConsumerConfig
 import org.apache.kafka.clients.producer.ProducerConfig
+
+import java.io.Serializable
+import java.util.Properties
 
 class BaseJobConfig(val config: Config, val jobName: String) extends Serializable {
 
@@ -19,7 +19,6 @@ class BaseJobConfig(val config: Config, val jobName: String) extends Serializabl
   val restartAttempts: Int = config.getInt("task.restart-strategy.attempts")
   val delayBetweenAttempts: Long = config.getLong("task.restart-strategy.delay")
   val parallelism: Int = config.getInt("task.parallelism")
-  
 
   val kafkaConsumerParallelism: Int = config.getInt("task.consumer.parallelism")
   // Only for Tests
@@ -32,19 +31,20 @@ class BaseJobConfig(val config: Config, val jobName: String) extends Serializabl
   val enableDistributedCheckpointing: Option[Boolean] = if (config.hasPath("job.enable.distributed.checkpointing")) Option(config.getBoolean("job.enable.distributed.checkpointing")) else None
   val checkpointingBaseUrl: Option[String] = if (config.hasPath("job.statebackend.base.url")) Option(config.getString("job.statebackend.base.url")) else None
   // By default checkpointing timeout is 10 mins
-  val checkpointingTimeout: Long = if(config.hasPath("task.checkpointing.timeout")) config.getLong("task.checkpointing.timeout") else 600000L
+  val checkpointingTimeout: Long = if (config.hasPath("task.checkpointing.timeout")) config.getLong("task.checkpointing.timeout") else 600000L
 
   // LMS Cassandra DB Config
   val lmsDbHost: String = config.getString("lms-cassandra.host")
   val lmsDbPort: Int = config.getInt("lms-cassandra.port")
-
 
   def kafkaConsumerProperties: Properties = {
     val properties = new Properties()
     properties.setProperty("bootstrap.servers", kafkaBrokerServers)
     properties.setProperty("group.id", groupId)
     properties.setProperty(ConsumerConfig.ISOLATION_LEVEL_CONFIG, "read_committed")
-    kafkaAutoOffsetReset.map { properties.setProperty("auto.offset.reset", _) }
+    kafkaAutoOffsetReset.map {
+      properties.setProperty("auto.offset.reset", _)
+    }
     properties
   }
 
@@ -55,5 +55,17 @@ class BaseJobConfig(val config: Config, val jobName: String) extends Serializabl
     properties.put(ProducerConfig.BATCH_SIZE_CONFIG, new Integer(16384 * 4))
     properties.put(ProducerConfig.COMPRESSION_TYPE_CONFIG, "snappy")
     properties
+  }
+
+  def getString(key: String, default: String): String = {
+    if (config.hasPath(key)) config.getString(key) else default
+  }
+
+  def getInt(key: String, default: Int): Int = {
+    if (config.hasPath(key)) config.getInt(key) else default
+  }
+
+  def getBoolean(key: String, default: Boolean): Boolean = {
+    if (config.hasPath(key)) config.getBoolean(key) else default
   }
 }
