+++ publish-pipeline/content-publish/src/main/scala/org/sunbird/job/content/publish/helpers/ECMLExtractor.scala
@@ -1,7 +1,7 @@
 package org.sunbird.job.content.publish.helpers
 
 import org.sunbird.job.content.publish.processor.{BaseProcessor, MissingAssetValidatorProcessor}
-import org.sunbird.job.publish.util.CloudStorageUtil
+import org.sunbird.job.util.CloudStorageUtil
 
 class ECMLExtractor(basePath: String, identifier: String)(implicit cloudStorageUtil: CloudStorageUtil) extends BaseProcessor(basePath, identifier) with MissingAssetValidatorProcessor {
