+++ publish-pipeline/content-publish/src/main/scala/org/sunbird/job/content/publish/processor/BaseProcessor.scala
@@ -1,9 +1,9 @@
 package org.sunbird.job.content.publish.processor
 
-import org.sunbird.job.publish.util.CloudStorageUtil
+import org.sunbird.job.util.CloudStorageUtil
 
 class BaseProcessor(basePath: String, identifier: String)(implicit cloudStorageUtil: CloudStorageUtil) extends IProcessor(basePath, identifier) {
-    override def process(ecrf: Plugin): Plugin = {
-        ecrf
-    }
+  override def process(ecrf: Plugin): Plugin = {
+    ecrf
+  }
 }
