+++ publish-pipeline/content-publish/src/main/scala/org/sunbird/job/content/publish/processor/IProcessor.scala
@@ -1,6 +1,6 @@
 package org.sunbird.job.content.publish.processor
 
-import org.sunbird.job.publish.util.CloudStorageUtil
+import org.sunbird.job.util.CloudStorageUtil
 
 abstract class IProcessor(basePath: String, identifier: String)(implicit cloudStorageUtil: CloudStorageUtil) {
