+++ publish-pipeline/publish-core/src/main/scala/org/sunbird/job/publish/config/PublishConfig.scala
@@ -5,19 +5,6 @@ import org.sunbird.job.BaseJobConfig
 
 class PublishConfig(override val config: Config, override val jobName: String) extends BaseJobConfig(config, jobName) {
 
-	def getString(key: String, default: String): String = {
-		if(config.hasPath(key)) config.getString(key) else default
-	}
-
-	def getInt(key: String, default: Int): Int = {
-		if(config.hasPath(key)) config.getInt(key) else default
-	}
-
-
-	def getBoolean(key: String, default: Boolean): Boolean = {
-		if(config.hasPath(key)) config.getBoolean(key) else default
-	}
-
-	def getConfig() = config
+  def getConfig() = config
 
 }
