+++ publish-pipeline/publish-core/src/test/scala/org/sunbird/job/publish/spec/EcarGeneratorSpec.scala
@@ -3,11 +3,10 @@ package org.sunbird.job.publish.spec
 import com.typesafe.config.{Config, ConfigFactory}
 import org.scalatest.{BeforeAndAfterAll, FlatSpec, Matchers}
 import org.sunbird.job.domain.`object`.DefinitionCache
-import org.sunbird.job.util.ScalaJsonUtil
+import org.sunbird.job.util.{CloudStorageUtil, ScalaJsonUtil}
 import org.sunbird.job.publish.config.PublishConfig
 import org.sunbird.job.publish.core.{DefinitionConfig, ObjectData}
 import org.sunbird.job.publish.helpers.EcarGenerator
-import org.sunbird.job.publish.util.CloudStorageUtil
 
 import scala.collection.JavaConverters._
 import scala.concurrent.ExecutionContext.Implicits.global
