+++ publish-pipeline/questionset-publish/src/main/scala/org/sunbird/job/questionset/publish/util/QuestionPublishUtil.scala
@@ -5,9 +5,8 @@ import org.sunbird.job.domain.`object`.DefinitionCache
 import org.sunbird.job.publish.config.PublishConfig
 import org.sunbird.job.publish.core.{DefinitionConfig, ExtDataConfig, ObjectData}
 import org.sunbird.job.publish.helpers.EcarPackageType
-import org.sunbird.job.publish.util.CloudStorageUtil
 import org.sunbird.job.questionset.publish.helpers.QuestionPublisher
-import org.sunbird.job.util.{CassandraUtil, HttpUtil, Neo4JUtil}
+import org.sunbird.job.util.{CassandraUtil, CloudStorageUtil, HttpUtil, Neo4JUtil}
 
 import scala.concurrent.ExecutionContext
 
@@ -21,7 +20,7 @@ object QuestionPublishUtil extends QuestionPublisher {
     logger.info("QuestionPublishUtil :::: publishing child question for questionset : " + identifier)
     objList.map(qData => {
       logger.info("QuestionPublishUtil :::: publishing child question : " + qData.identifier)
-      val obj = getObject(qData.identifier, qData.pkgVersion, qData.mimeType, qData.metadata.getOrElse("publish_type","Public").toString, readerConfig)(neo4JUtil, cassandraUtil)
+      val obj = getObject(qData.identifier, qData.pkgVersion, qData.mimeType, qData.metadata.getOrElse("publish_type", "Public").toString, readerConfig)(neo4JUtil, cassandraUtil)
       val messages: List[String] = validate(obj, obj.identifier, validateQuestion)
       if (messages.isEmpty) {
         val enrichedObj = enrichObject(obj)(neo4JUtil, cassandraUtil, readerConfig, cloudStorageUtil, config, definitionCache, definitionConfig)
